// src/domains/inventory/dto/update-product.dto.ts
import type { ProductStatus } from "@/generated/prisma/enums";

export interface UpdateProductDto {
  sku?: string;
  name?: string;

  description?: string;

  brandId?: string;
  categoryId?: string;
  supplierId?: string;

  barcode?: string;
  unit?: string;

  costPrice?: number;
  sellingPrice?: number;

  currentStock?: number;
  minimumStock?: number;

  status?: ProductStatus;
}